# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LandOverSEA - International Dating & Cultural Connection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /safety/users/resolve/{supabaseUserId}:
    get:
      operationId: resolveSafetyUser
      tags: [safety]
      summary: Resolve a linked Supabase user for safety actions
      parameters:
        - name: supabaseUserId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      responses:
        "200":
          description: Linked local safety identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafetyUserResolution"
        "400":
          description: Invalid Supabase user ID
        "404":
          description: Supabase user has no linked local safety identity
  /safety/block:
    post:
      operationId: blockUser
      tags: [safety]
      summary: Block a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid target or self-block
        "404":
          description: Target user not found
  /safety/block/{blockedUserId}:
    delete:
      operationId: unblockUser
      tags: [safety]
      summary: Unblock a user
      parameters:
        - name: blockedUserId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid target or self-unblock
        "404":
          description: Target user not found
...
  /safety/report:
    post:
      operationId: reportUser
      tags: [safety]
      summary: Report a user or content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── AUTH ──────────────────────────────────────────────────────────────
...
    BlockInput:
      type: object
      required: [blockedUserId]
      properties:
        blockedUserId:
          type: number
        reason:
          type: ["string", "null"]
...
    ReportInput:
      type: object
      required: [reportedUserId, reason]
      properties:
        reportedUserId:
          type: number
        reason:
          type: string
        description:
          type: ["string", "null"]
        messageId:
          type: ["number", "null"]
    SafetyUserResolution:
      type: object
      required: [userId]
      properties:
        userId:
          type: number
...